RavenDB-24836 Address PR #23363 review feedback
- Split TimeSeriesRangePicker.tsx into TimePicker.tsx and RangeDatePickerHeader.tsx
- Make date presets timezone-aware so "Today" etc. resolve in the selected zone
- Share one time-parse format list between buildInstant and parseTimeParts, and
  normalise partial time entry on blur
- Preserve typed/prefilled values when switching Between/Before/After instead of
  discarding them via an unconditional preset reset
- Fix stale comment about picker timezone not affecting the grid
- Replace DeleteTimeSeriesRangeModal's manual delete state with useAsyncCallback
- Use project sass/less tokens instead of raw bootstrap vars and magic numbers
