+++ src/Raven.Studio/wwwroot/Content/css/pages/time-series-edit.less
@@ -1,6 +1,6 @@
 .edit-time-series {
     .document-header {
-        margin-bottom: 30px;
+        margin-bottom: @grid-gutter-width;
 
         .dropdown-toggle {
             min-width: 120px;
